Improves Steam User Plus badgeList function
At my first implementation, badgeList was only collecting information related to games that still had card drops.
Thinking about it later, I found out that it didn't make sense, and deciding to work with a subset of the badgelist should be something to be handled by the function caller.
This commit updates the badgeList function so it now retrieves the entire list of games, it also updates the main index script so it filters the list before idling, so it works only with the ones that still have card drops.
As a side note, foil badges in the badge list are skipped, because they don't have "time_played" information, I'm still unsure if I should skip these, but it will stay this way for now.

I'm also using this commit to add some temp files that I create for testing to the .gitignore.

Signed-off-by: gabrieldrs <gabriel@sezefredo.com.br>