Web App now shows fail/success according to API outcome
At my first implementation, the success/fail message were based on whether the user had typed their credentials or not, and had no actual check on the API backed. It's now waiting for the Steam API response in order to give the correct message, it also takes the reason for the failure and displays on the front-end.

I know the way it's implemented is really rudimentary, maybe one path to follow from now on would be to improve the front-end so it's more application like.

I have been thinking about using Electron on the future, making it easier to install/use locally.

Signed-off-by: gabrieldrs <gabriel@sezefredo.com.br>