Adds support to email token authentication
When a user in steam has not configred two factor authentication in their mobile phone, steam uses their email to send the confirmation token.
In this case, the confirmation token has to be sent to steam in another json field during the logon step.
This commit identifies the type of error that occurs in a non successful logon attempt, and use that to decide if it will send the token as a two-factor authentication or an email-code authentication.

This solves bug reported at #6

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>