Removal of Steam User Plus code from the repository
Previously I had implemented a specific steam user plus lib into this repo, but as the project grew, I think this is a module by itself.
By moving it to a module, it's easier to define responsabilities and also easier to maintain. Improvements on the steam user object will be handled by the new module 'steam-user-plus'.
I created a new function in the index.js so the caching logic stays here (where it makes more sense).

I also cleaned package.js removing unnecessary packages that were only being used in the steam-user-plus code.

Signed-off-by: Gabriel R. Sezefredo <gabriel@sezefredo.com.br>